+++ src/Ajax.php
@@ -3,7 +3,7 @@
 namespace TrustedLogin;
 
 // Exit if accessed directly
-if ( ! defined('ABSPATH') ) {
+if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
@@ -23,6 +23,19 @@ final class Ajax {
 	 * @var null|\TrustedLogin\Logging $logging
 	 */
 	private $logging;
+
+	/**
+	 * @var string[] Fields that may be included in the support data.
+	 * @see grantAccess() in trustedlogin.js
+	 */
+	private $generate_support_fields = array(
+		'action',
+		'vendor',
+		'_nonce',
+		'reference_id',
+		'debug_data_consent',
+		'ticket',
+	);
 
 	/**
 	 * Cron constructor.
@@ -51,7 +64,10 @@ final class Ajax {
 	 */
 	public function ajax_generate_support() {
 
-		if ( empty( $_POST['vendor'] ) ) {
+		// Remove any fields that are not in the $ajax_fields array.
+		$posted_data = array_intersect_key( $_POST, array_flip( $this->generate_support_fields ) );
+
+		if ( empty( $posted_data['vendor'] ) ) {
 
 			$this->logging->log( 'Vendor not defined in TrustedLogin configuration.', __METHOD__, 'critical' );
 
@@ -60,34 +76,39 @@ final class Ajax {
 
 		// There are multiple TrustedLogin instances, and this is not the one being called.
 		// This should not occur, since the AJAX action is namespaced.
-		if ( $this->config->ns() !== $_POST['vendor'] ) {
+		if ( $this->config->ns() !== $posted_data['vendor'] ) {
 
 			$this->logging->log( 'Vendor does not match TrustedLogin configuration.', __METHOD__, 'critical' );
 
 			wp_send_json_error( array( 'message' => 'Vendor does not match.' ) );
+
 			return;
 		}
 
-		if ( empty( $_POST['_nonce'] ) ) {
+		if ( empty( $posted_data['_nonce'] ) ) {
 			wp_send_json_error( array( 'message' => 'Nonce not sent in the request.' ) );
 		}
 
 		if ( ! check_ajax_referer( 'tl_nonce-' . get_current_user_id(), '_nonce', false ) ) {
-			wp_send_json_error( array( 'message' => 'Verification issue: Request could not be verified. Please reload the page.' ) );
+			wp_send_json_error( array( 'message' => esc_html__( 'Verification issue: Request could not be verified. Please reload the page.', 'trustedlogin' ) ) );
 		}
 
 		if ( ! current_user_can( 'create_users' ) ) {
 
 			$this->logging->log( 'Current user does not have `create_users` capability when trying to grant access.', __METHOD__, 'error' );
 
-			wp_send_json_error( array( 'message' => 'You do not have the ability to create users.' ) );
+			wp_send_json_error( array( 'message' => esc_html__( 'You do not have the ability to create users.', 'trustedlogin' ) ) );
 		}
 
 		$client = new Client( $this->config, false );
 
-		$include_debug_data = ! empty( $_POST['debug_data_consent'] );
+		// Passed from grantAccess() in trustedlogin.js.
+		$include_debug_data = ! empty( $posted_data['debug_data_consent'] );
 
-		$response = $client->grant_access( $include_debug_data );
+		// Passed from grantAccess() in trustedlogin.js.
+		$ticket_data = ! empty( $posted_data['ticket'] ) ? $posted_data['ticket'] : null;
+
+		$response = $client->grant_access( $include_debug_data, $ticket_data );
 
 		if ( is_wp_error( $response ) ) {