+++ src/Client.php
@@ -9,7 +9,7 @@
  *
  * 0. If you haven't already, sign up for a TrustedLogin account {@see https://www.trustedlogin.com}
  * 1. Namespace the installation ({@see https://www.trustedlogin.com/configuration/} to learn how)
- * 2. Instantiate this class with a configuration array (really, {@see https://www.trustedlogin.com/configuration/} for more info)
+ * 2. Instantiate this class with a configuration object (really, go see {@see https://www.trustedlogin.com/configuration/} for more info)
  *
  * Class Client
  *
@@ -17,10 +17,11 @@
  *
  * @copyright 2021 Katz Web Services, Inc.
  */
+
 namespace TrustedLogin;
 
 // Exit if accessed directly
-if ( ! defined('ABSPATH') ) {
+if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
@@ -36,7 +37,7 @@ final class Client {
 	 * @var string The current SDK version.
 	 * @since 1.0.0
 	 */
-	const VERSION = '1.4.0';
+	const VERSION = '1.5.0';
 
 	/**
 	 * @var Config
@@ -88,6 +89,7 @@ final class Client {
 	 */
 	private $site_access;
 
+
 	/**
 	 * TrustedLogin constructor.
 	 *
@@ -125,13 +127,16 @@ final class Client {
 
 		$this->support_user = new SupportUser( $this->config, $this->logging );
 
-		$this->admin = new Admin( $this->config, $this->logging );
+		$this->site_access = new SiteAccess( $this->config, $this->logging );
+
+		$form = new Form( $this->config, $this->logging, $this->support_user, $this->site_access );
+
+		$this->admin = new Admin( $this->config, $form, $this->support_user );
 
 		$this->ajax = new Ajax( $this->config, $this->logging );
 
 		$this->remote = new Remote( $this->config, $this->logging );
 
-		$this->site_access = new SiteAccess( $this->config, $this->logging );
 
 		if ( $init ) {
 			$this->init();
@@ -161,7 +166,7 @@ final class Client {
 
 		// Disables namespaced client if `TRUSTEDLOGIN_DISABLE_{NS}` is defined and truthy.
 		if ( defined( 'TRUSTEDLOGIN_DISABLE_' . strtoupper( $ns ) ) && constant( 'TRUSTEDLOGIN_DISABLE_' . strtoupper( $ns ) ) ) {
-			return new WP_Error( 'disabled_for_namespace', 'TrustedLogin has been disabled for this namespace using the TRUSTEDLOGIN_DISABLE_' . $ns .' constant.' );
+			return new WP_Error( 'disabled_for_namespace', 'TrustedLogin has been disabled for this namespace using the TRUSTEDLOGIN_DISABLE_' . $ns . ' constant.' );
 		}
 
 		$meets_requirements = Encryption::meets_requirements();
@@ -204,9 +209,14 @@ final class Client {
 	/**
 	 * This creates a TrustedLogin user ✨
 	 *
+	 * @since 1.5.0 Added $ticket_data parameter.
+	 *
+	 * @param bool $include_debug_data Whether to include debug data in the response.
+	 * @param array|null $ticket_data If provided, customer-provided data associated with the access request.
+	 *
 	 * @return array|WP_Error
 	 */
-	public function grant_access( $include_debug_data = false ) {
+	public function grant_access( $include_debug_data = false, $ticket_data = null ) {
 
 		if ( ! self::$valid_config ) {
 			return new \WP_Error( 'invalid_configuration', 'TrustedLogin has not been properly configured or instantiated.', array( 'error_code' => 424 ) );
@@ -365,6 +375,10 @@ final class Client {
 			$action_data['debug_data'] = $this->get_debug_data();
 		}
 
+		if ( $ticket_data ) {
+			$action_data['ticket'] = $ticket_data;
+		}
+
 		/**
 		 * @usedby Remote::maybe_send_webhook()
 		 */
@@ -438,11 +452,17 @@ final class Client {
 
 		try {
 
-			add_filter( 'trustedlogin/' . $this->config->ns() . '/envelope/meta', array( $this, 'add_meta_to_envelope' ) );
+			add_filter( 'trustedlogin/' . $this->config->ns() . '/envelope/meta', array(
+				$this,
+				'add_meta_to_envelope'
+			) );
 
 			$updated = $this->site_access->sync_secret( $secret_id, $site_identifier_hash, 'extend' );
 
-			remove_filter( 'trustedlogin/' . $this->config->ns() . '/envelope/meta', array( $this, 'add_meta_to_envelope' ) );
+			remove_filter( 'trustedlogin/' . $this->config->ns() . '/envelope/meta', array(
+				$this,
+				'add_meta_to_envelope'
+			) );
 
 		} catch ( Exception $e ) {
 
@@ -472,10 +492,10 @@ final class Client {
 		 * @usedby Remote::maybe_send_webhook()
 		 */
 		do_action( 'trustedlogin/' . $this->config->ns() . '/access/extended', array(
-			'url'    => get_site_url(),
-			'ns' => $this->config->ns(),
-			'action' => 'extended',
-			'ref' => self::get_reference_id(),
+			'url'        => get_site_url(),
+			'ns'         => $this->config->ns(),
+			'action'     => 'extended',
+			'ref'        => self::get_reference_id(),
 			'access_key' => $this->site_access->get_access_key(),
 		) );
 
@@ -515,8 +535,8 @@ final class Client {
 		}
 
 		$site_identifier_hash = $this->support_user->get_site_hash( $user );
-		$endpoint_hash = $this->endpoint->get_hash( $site_identifier_hash );
-		$secret_id = $this->endpoint->generate_secret_id( $site_identifier_hash, $endpoint_hash );
+		$endpoint_hash        = $this->endpoint->get_hash( $site_identifier_hash );
+		$secret_id            = $this->endpoint->generate_secret_id( $site_identifier_hash, $endpoint_hash );
 
 		// Revoke site in SaaS
 		$site_revoked = $this->site_access->revoke( $secret_id, $this->remote );
@@ -540,6 +560,7 @@ final class Client {
 
 		if ( ! empty( $should_be_deleted ) ) {
 			$this->logging->log( 'User #' . $should_be_deleted->ID . ' was not removed', __METHOD__, 'error' );
+
 			return new \WP_Error( 'support_user_not_deleted', esc_html__( 'The support user was not deleted.', 'trustedlogin' ) );
 		}
 
@@ -548,7 +569,7 @@ final class Client {
 		 */
 		do_action( 'trustedlogin/' . $this->config->ns() . '/access/revoked', array(
 			'url'    => get_site_url(),
-			'ns' => $this->config->ns(),
+			'ns'     => $this->config->ns(),
 			'action' => 'revoked',
 		) );